Added checks for change in pos & rot
This way it reduces the number of sends to the server by a huge factor.
Before, it was sending ~60 per second easily reaching over a thousand
calls in 20 seconds of play time * the number of players.  Now, it'll
only send the information update on a change.

Also, finished video 3.  Now for video 4.
https://www.youtube.com/watch?v=CdbzZSHPN4E